tests/visual/pyproject.toml
[project]
name = "circuits-visual-tests"
version = "0.1.0"
requires-python = ">=3.12"
description = "Playwright-driven visual/layout/perf regression tests for circuits.com"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-playwright>=0.5",
]

[tool.pytest.ini_options]
testpaths = ["."]
addopts = "-v --browser chromium"
